containments/desktop: fix single-click to rename
FolderView has a feature that allows the user to single-click the label
of an already selected item to trigger renaming that item (given a
particular combination of settings), mirroring similar functionality in
Dolphin. This is, again as in Dolphin, supposed to not apply to cases
where the label is double-clicked, which should open the double-clicked
file. However, the detection of double-clicks does not work: it relies
on the doubleClickInProgress property, which is set later in the event
handler (and never reached in this case as the handler returns early),
and so is almost always false (unless the clicked element changes mid-
double-click).

This change modifies the event handling logic to instead use the regular
logic to wait for a double-click, and triggers the renaming function
from the timer if applicable.

BUG: 493124
FIXED-IN: 6.2


(cherry picked from commit 36e0d2dd4198b7948adfb55d82c96852bba167de)

Co-authored-by: Christoph Wolk <christoph.wolk@gmx.de>