applets/kickoff: close on session/power action
The power and session actions currently do not explicitly close kickoff
like starting an application does. In most cases this has no consequence
as kickoff will be closed by the session/power action itself or its
consequences. There are some, though, that do not necessarily do so; one
of them is sleep if the computer is configured to not lock the screen on
waking (otherwise locking would close kickoff).

This change makes the power/session buttons behave like regular entries
and close the popup when activated. The user can still prevent this by
pinning kickoff.

BUG: 493439