applets/taskmanager: choose better badge size
The badge size on the task manager is based on the height of the task
manager button; more specifically it is set to have a height of 40 % of
the button's height. This generally works very well, but fails in some
configurations on vertical panels with IOTM: if the icon spacing is
changed, it will affect the height rather than the width of the button
(as with a horizontal IOTM), and thus also scale the badge, making the
label unreasonably large with large spacing. On a very thin, small
spacing panel on the other hand, the label may become unreadably tiny.

This change makes the badge size depend on the actual icon size instead.
As the icon is always a little smaller than the task manager button, the
scaling factor is slightly increased (to 45%) to keep badge size on the
default panel the same as it is now.