FolderViewLayer.qml: Load the icon positions when geometry changes
We need to load the icon positions when screen geometry changes,
otherwise we risk using wrong geometry when screen is turned off or
otherwise changed, then saving the icon positions of that geometry.

For example, when turning off screen in single-monitor setup,
the resolution will change to 0x0. This then moves the icons around.
Then opening the screen doesn't move the icons, so the icons stay in the
0x0 positions, and save them to the "normal" resolution.

This fixes it by loading the positions whenever the geometries change

BUG:493569