applets/pager: Remove animations
The pager applet provides a miniature of the virtual desktop grid and by
default, it's placed in the panel. On the other hand, the main purpose of
animations is to notify or guide the user through state changes. Due to
the location and the size of the pager, the user focus will likely be
elsewhere and these animations will go unnoticed.

Unfortunately, these animations have caused some issues too. One was with
drag and drop. 463b6798aecfbe81e58dae2c23a0449d6537e581 attempted to
address that by dropping x and y animations. That change should have
probably dropped width and height animations too because they alone don't
serve the real purpose. Another issue is caused by flaws in the animation
driving code in QtQuick, which can result in plasmashell triggering
constant repaints until the panel that contains the pager is shown again.

So, given the aforementioned issues and small net benefit of these
animations, this change drops animations in the pager. It is not the
definite fix for upstream issues but it should at least help us mitigate
repaint issues that can be observed on setups with the current defaults.

CCBUG: 487895


(cherry picked from commit 8f2b64e329301b958413ac64bd64ed22d8110e3a)

Co-authored-by: Vlad Zahorodnii <vlad.zahorodnii@kde.org>