kcms/tablet: Improve calibration accuracy
The matrix we get from the calibration process should be multipled
against the existing calibration matrix from the device, since it was
used to create it in the first place. This also makes the refinement
feature actually useful, because you can keep going and possibly get a
more accurate calibration.