Foldermodel: Save item positions when any item gets renamed
Previously we did not save the item position at all after renaming, so
if user would rename a file and do nothing else, it's position would not
get saved. Then it would lose its position if screen is turned off etc.

This adds a new signal FolderModel::itemRenamed, that is emitted every
time an item gets renamed (with singleshot connection).

In positioner we connect to this signal with the updatePosition slot.