positioner: Make sure we check deferMovePositions exist
Checking `m_deferMovePositions.count()` seems to be always true,
even when its empty.

Make sure we check the size is more than 0 so we don't move items
around without any reason to do so.

We also do not emit `positionsChanged` if we have not even touched the positions,
to avoid deleting old configuration

CCBUG:493569