FolderViewLayer: Ignore icon positioning if screen is off or being resized
Due to panels being added to the screen, the actual space for icons is
in a flux until the resizing is done. This patch adds a timer
that is being restarted every time the screen sizes change,
and when the sizes stop changing for one second, we apply any of the
positions and allow saving positions again.

This patch also checks if there is a screen with Window.window.
If Window.window is null, there is no screen, so there is no reason
for the icons to be positioned or their positions to be saved, since
this would overwrite any settings we have when screen is on.

CCBUG:493569