emojier: quit on copy
After you select an emoji, what's the thing you're going to want to do
next? Paste it! But to do this, you have to get rid of the window.
That's a manual step that slows you down.

Instead, let's quit automatically after copying an emoji. To prevent
disorientation, it shows a system notification indicating that the emoji
has been copied to the clipboard.

BUG: 467662
FIXED-IN: 6.8.0