applets/taskmanager: At least show something when thumbnails are disabled
There's an option to not show window thumbnails, but when you use it,
all tooltips for tasks disappear, not just window thumbnails. This isn't
ideal since you lose all feedback about the hovered task. In addition,
the behavior doesn't match the text of the option, since more than just
tooltips disappear.

Instead, show the small title-and-subtitle style tooltips.

BUG: 492530
FIXED-IN: 6.3.0