positioner: Load position config after deferred moves
When deferred moves are done for all the items, we should load the
configuration afterwards to make sure any items on the configuration
are placed in the correct place.

Deferred moves often happen during folder view margin resizing or after
folderitems loading/changing. Thus they can sometimes load unexpected
positions (because they are looking for the best possible position for
the item), even if the user has configured them to load in some specific
way.

BUG:496556