Only return valid task item size if task manager has been resized to fit panel
Previously the task manager would briefly have null width and height, before
being resized to fit the panel. However, the "preferredMaxWidth" for tasks
would still be positive, as it adds margins to it, and tasks would then
assume a positive width. When the proper values are set, this casuses
an extra resize animation.

CCBUG:447476