kcms/tablet: Overhaul UI/UX
The current KCM is too simply-designed to handle future features
(pressure curve, more configuration options, etc) and it's also kind of
struggling with what it has today.

So this splits it up into three tabs: "Display", "Pen", and "Pad"
which will give us plenty of room in the future to place more features
and options. Some tabs are hidden depending on what the device
supports.