positioner: move icon to creation spot on create
If screen is in use, we can safely assume this action was by user when
they create a new file on desktop. We can then save the
location and then load the positions to make sure the item
moves in correct place.

See https://invent.kde.org/plasma/plasma-desktop/-/merge_requests/2607#note_1092235