kcms/tablet: Add support for configuring tablet dials
This exposes the new rebinding feature in KWin for these input devices,
which are basically mouse scroll wheels stuck to the sides of a graphics
tablet. These are typically used for zooming in & out of a canvas, or
changing your brush size.

Two new rebind modes are introduced: scroll wheels and up/down keybinds.
These are exclusive to devices that only have relative up/down input
like these dials.