kcms/touchpad: Use RadioSelector from kirigami-addons
The "Pointer acceleration" and "Scrolling" selectors are the ideal
use case for this component, given a fixed small number of choices
(two for both, with pointer acceleration getting a third one soon)
and short option texts.

Furthermore, switching to a horizontal layout for the "Scrolling"
radio buttons makes the "Invert scroll direction" checkbox underneath
less awkward.

I'm not quite convinced that the current styling really fits in,
though.