applets/kickoff: fix KickoffGridView runtime warning
The commit for optional activation on hover (05a4d828) introduced a
check for the mainContentView property, and a reference to the Plasmoid
configuration in KickoffGridView. This, however, is not a property of
this component, only of KickoffListView. The component also does not
import org.kde.plasma.plasmoid and thus cannot access Plasmoid. This
causes runtime warnings when the Kickoff applet is loaded.

This change sets hoverEnabled to always be true, rather than depending
on mainContentView status or configuration, as KickoffGridView is always
a mainContentView.