Reduce clutter on global edit mode toolbar
The global edit mode toolbar right now as a lot of stuff on it, some of
which is not really very relevant to editing the desktop and panels or
isn't important to have here:

1. "Global Themes" doesn't have anything to do with the desktop and
   panels.
2. "Display Configuration" doesn't have anything to to with the desktop
   and panels, and is also visible in the menu used to enter edit mode
   in the first place.
3. "Wallpaper and desktop" definitely does have something to do with the
   desktop and panels, but it's not only visible in the menu used to
   enter edit mode in the first place, and also the wallpaper component
   is also visible in System Settings now, making access from here less
   unnecessary.

If you manage to enter edit mode using a touchscreen, the full contents
of the desktop context menu are available via a hamburger button on the
global edit mode toolbar as well.

As such, these items are mostly clutter on the toolbar, so let's remove
them to clean up the presentation of what matters most.