applets/taskmanager: make no-image tooltip clickable
355e5a88 enabled basic tooltips when image previews are disabled. These
"basic" tooltips still have most of the  functions of the regular
toolips with previews; they show each window of the task, allow users to
close these windows, access media controls, etc. What's absent is maybe
the most important function: focusing one of those windows. This is
because that is done by clicking the thumbnail, and the thumbnail-less
tooltips have no alternate method. This is unfortunate, and confusing as
the often large areas look enticingly clickable.

This change makes the header clickable in case there are no thumbnails;
the version with thumbnails should be unaffected.