applets/taskmanager: only show media controls once
The media/volume controls in the tooltips are shown for each window in
the group. This makes it look like they would be able to control each
item separately, but they do not - each of them is simply a copy of the
others and applies to the same instance of the media player (usually the
one registered first) and the global application volume. This is
wasteful, and rather confusing.

This change makes it so only one item displays media and volume
controls; in general this will be the first item as this is likely the
one that is actually being controlled. If album art is set and would be
shown for a window, it will instead be displayed for that one.