containments/desktop: properly map position to find the mouseover item
GridView.itemAt() is documented to accept coordinated in *contentItem*
coordinates. passing coordinates in view coordinates will cause items to
not
be found if contentItem position is not 0,0

With right align, the contentItem is aligned to the right edge of the
view and items are positioned from there at negative x coordinates.

BUG:497498


(cherry picked from commit 509b655ef2edef04a13d9afaab6349911beaaeb1)

509b655e containments/desktop: properly map position to find the mouseover item

Co-authored-by: Marco Martin <notmart@gmail.com>