applets/folder: Don't use DesktopSchemeHelper::getFileUrl to get title
It isn't needed since we don't directly display it to the user; instead
we poke and prod it in various ways and then return the processed
string. This means we need the original URL, not a modified version of
it, which makes KFilePlacesModel::closestItem() return the wrong thing.

KFilePlacesModel::closestItem() isn't doing anything technically
incorrect here; it's our problem for giving it malformed data.

This fixes the Folder View widget having the wrong title when showing
the contents of desktop:/, as it does by default.

BUG: 494430
FIXED-IN: 6.3.0