Revert "applets/kickoff: change category at most once per frame"
This reverts commit d5b9b7991a7e262bc34bf46befb56cfc4f029f51.
The functionality is so broken as I question if this was even
tested at all. This breaks changing the index with keys as
currentIndex is always reset to pendingIndex and for the
same reason by clicking. Resulting in being unable to
change categories at all. Even after fixing this the design
doesn't work, as not changing index doesn't cause a redraw
which means no frameSwapped occurs.
Further the CI pipeline failing on kickofftest was ignored
when merging.