Rewrite timezone section of datetime kcm in QML
This revamps the timezone selection page, which will now show a map with all available
("region/location") timezones, _and_ a combobox to select a specific timezone (even
those that might not be on the map, such as "UTC+04").

The new approach is:
- Offline tiles are used to display the map, through QML's MapView. The map tiles are
  directly taken from openstreetmap's website.
- Timezone data is taken from the comprehensive list from 

This allows to see and select actual geographic timezones by clicking on them, without
requiring any third party API during usage or any internet connection at all. It still
allows to select timezones that don't directly map to a region using the combobox.

BUG: 487873
FIXED-IN: 6.3.0