applets/kickoff: make top corners and left/right header paddings perfect
Right now these are imperfect due to fancy math on PlasmoidHeading that
it isn't compatible with. Switch to a simpler implementation that allows
us to perfect the top corners and the left and right paddings of the
toolbar items.

BUG: 473061
FIXED-IN: 6.3.0


(cherry picked from commit 8df770185237545b5f6a2e247290fd95e4fd3d36)

Co-authored-by: Nate Graham <nate@kde.org>