applets/taskmanager: make vertical tooltip compact
The task tooltips when shown in vertical orientation (e.g. side panel)
have quite a bit of wasted space: the App name for all windows
in a group except the first, and empty window titles resulting from a
matching app name or a media player. This makes the tooltips larger and
more obtrusive than they need to be.

This change removes those lines instead of leaving them empty. This
applies only to vertical tooltips, as all windows in a group should have
the same height.