applets/taskmanager: avoid close button near panel
The task tooltips always have the order text>(badge)>close button. On a
right screen edge panel in LtR direction, or a left screen edge panel in
RTL direction, this means that the close button is right next to the
task icon, in a place that is often dangerously close to a convenient
place to click to activate the window. This is the case on tooltips with
thumbnails disabled, which are fully clickable and very small.

This change makes the close button dependent on panel edge and layout
direction, such that it will always show up on the side of the tooltip
opposite to the panel.