kcms/workspaceoptions: improve helper text position
Some radiobuttons and checkboxes have helper text displayed in the UI.
This text is indented a bit, but is not properly aligned to the text
of the widget it belongs to, so its position feels arbitrary. A similar
issue existed in the touchpad kcm and was fixed in
f39388c64c9727c6ecfbe71ef01b1fab94d40f38.

This change aligns the labels properly, leading to a cleaner visual
experience.