kcms/workspaceoptions: screen-read helper text
Some options have helper text displayed below the respective control.
This information is only presented visually, it is not automatically
read out by screen readers as the user moves through the form.

This change sets the helper text as an Accessible.description for all
controls that have a helper text.