applets/taskmanager: prevent empty tooltip items
For grouped tasks, the tooltip instance for a particular window can
be completely empty (save for the close button) under some very narrow
circumstances, namely the app name and window title have to be
identical, the window has to be the second or later, there has to be no
applicable subtitle (like virtual desktop number) and thumbnails have to
be disabled. In those cases, there is nothing really useful to show, but
having absolutely nothing seems even more broken.

This change makes the window title show even if it is a duplicate of the
app name, but only if the tooltip entry for that window would otherwise
be completely empty. As the conditions for what to show and when are
getting hard to read, they were also refactored to make them more
descriptive.


(cherry picked from commit 47186fdd53459be996771339646e30fad232b52e)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>