kcms/touchpad: respect layoutDirection
KCMUtils has some magic that makes loaded KQuickConfigModules
automatically respect layoutDirection. kcm_touchpad has a qml UI, but
is not implemented a a KQuickConfigModule, but as a KConfigModule that
manually loads qml files. Thus, the automagic does not happen, and
kcm_touchpad will not reflect the locale's reading direction nor the
--reverse flag.

This change explicitly sets the relevant LayoutMirroring properties.