Desktop: Make context menu for trash link and contents more contextual
There's already some special handling for the context menu of the trash
link, but it's currently not contextual enough: the menu shows many
items that are either not relevant to the trash, or expose the
implementation detail that the item is actually a .desktop file.

The same is true for items *in* the trash.

Expand on this to clean up their menus even more.

BUG: 416954
FIXED-IN: 6.4.0