applets/kicker: respect layoutDirection in keynav
Kicker's arrow key navigation does not follow layoutdirection, so
in a RtL locale, pressing left arrow will move to the right kicker
column, and to open a submenu the user needs to press the arrow key in
the opposite direction from the one indicated by the submenu indicator
arrow.

This change makes Kicker's arrow key handling adjust to layoutDirection.
A small bug that prevents wrapping from the search field to the opposite
end of the columns is also fixed.