desktop: Replace chord keyboard shortcut with something comprehensible
The desktop currently has a local-scoped shortcut to open the config
window. This is odd, but because the desktop isn't global (there can be
multiple desktops), in principle is can make sense for keyboard
accessibility.

Unfortunately it uses a chord shortcut, which users don't understand.
Replace it with a normal-style shortcut.

Part of https://invent.kde.org/plasma/plasma-workspace/-/issues/148.

CCBUG: 433292
CCBUG: 433841