applets/kicker: fix DashboardRepresentation warning
The ActionMenu component was changed to no longer inherit from Item.
While the use in cascading kicker was adapted to this, the one in
DashboardRepresentation was not. It stores the ActionMenu in an Item
property, and as the types do not match a warning is printed.

This change sets the type of the property to match what is actually
stored.