applets/kicker: make dash/grid keynav not broken in RtL
Arrow key navigation in dashboard/grid view is currently completely
broken in RtL, and not just in a "arrow keys go the wrong direction"
way, due to the complex way that multiple grids work and various
components taking RtL into account or not.

This change at least makes it function somewhat reasonably.