applets/taskmanager: fix TriangleMouseFiler in RtL
Taskmanager sets TriangleMouseFilter's width to the childrenRect's
width; this works well in general, but breaks on RtL. The childrenRect
is relative to the parent (the TMF), and its coordinates get
recalculated as the parent's coordinates change, which in the RtL case
is when each item is added, causing the width to get updated -> binding
loop. The width ends up getting stuck at most one task entry long, and
for the others the TriangleFilter does not work as its width does not
span them.

We already set the width of the taskList explicitly by calculating the
children's width for other purposes, so we can just reuse that and avoid
the loop.