lockscreen: Fade OSD just after timeout
The OSD hasn't been doing that long fadeout for a long time.
Adjust the lock screen to match and just hide it with a short
animation after the full timeout.

Also enable layer effect during the animation to avoid opacity
layering of the ProgressBar where you can see halfway through it.