kcms/touchpad: Rename setTouchpadOff() to setTouchpadSuspended()
Note that this is X11 only because the Wayland session does not
run the kded where this method is in use.

On an X11 property level, this is somewhat incorrect because this
`XlibTouchpad` function actually does use the "enabled" property.
Really it's the "SendEvents" property used in `LibinputTouchpad`
that's the "real" device suspend.

But from a user point of view, that doesn't matter. The semantics
of this function are to turn the touchpad on or off without applying
other device properties, without persisting anything, and "suspended"
reflects those semantics well.

The consequence is that on X11, this continues to stack on top of
the touchpad's actual `enabled` property, which is implemented using
libinput "SendEvents".