WidgetExplorer: improve screen reader accessibility
Widget Explorer contains a lot of information about widgets and plenty
of buttons, but screen reader use is rather inconsistent - much of that
is not read out, including the central aspects of the explorer: widget
names and descriptions.

This change makes the components from the visual interface accessible
to screen readers. Along the way, the text for uninstalling widgets was
adjusted a bit to make it clearer and more consistent, using "uninstall"
thoughout and leaving "remove" for the action that removes active
widgets from panels/desktop without uninstalling them.