applets/kicker: fix non-unique default highlight
When searching, kicker sets the currentItem to 0 as the number of hits
for a runner changes, but only if it is the first visible runner, to
indicate which entry will be activated if the user presses Enter from
the search field. The problem is that runners are asynchronous, and
lower-ranked runners may return first, at which point they are the first
visible column and have their first entry highlighted, only to be pushed
aside later as higher-ranked runners come in. As a consequence, multiple
runners may end up with a highlighted item.

Instead, update the selection status when any runner finishes.