applets/kicker: actually select delegate on arrow
When the user presses an arrow key from the search field, the selection
is moved to the second or final entry in the list, depending on the key
pressed. Then focus is passed to the column. This doesn't work properly
though, as the actual keyboard focus does not end up on the delegate
yet, only after the user presses another arrow. If the user tries to
activate the item highlighted after a single key press, the Return press
falls back to the search field which always activates the first entry in
the list.

Instead, expose the actual highlighted list item, then focus it directly
instead of focusing the column.

BUG: 499971
FIXED-IN: 6.3.1