applets/TaskManager: don't set non-existing property
MouseHandler sets animating to false onEntered. This makes sense for
dragging tasks in the actual task manager, but the type is also used
in GroupDialog. Here it's much less useful; instead of having an effect
it just prints a runtime warning because of the missing property.

Instead, only set animating to false if it's true (and thus not
undefined like a non-existing property). As animating is a bool, it
shouldn't ever be legitimately undefined.