applets/trash: allow to optionally be in the System Tray
It's disabled by default like the Disk Quota widget. But when using
"Shown when relevant" mode, it will be visible when it has anything
in it.

This provides a handy systemwide location to access and empty the trash
from, for people who prefer a Mac Dock style with a trash on it, but
with the enhancement that it only appears when non-empty.

BUG: 446548
BUG: 447709
FIXED-IN: 6.4.0