kcms/tablet: Fix appearance of input area if changing it is unsupported
This feature is only supported on libinput >=1.27, but I forgot to
re-test it using an older libinput and this bug slipped into 6.3.

The source of the bug isn't immediately obvious, this is chalked up to
intentional behavior in libkwindevices. If a property is reported as
unsupported, then it default constructs a value (in this case, a
QRectF.) But we don't want a default-constructed QRectF, we need a
QRectF with a size of 1x1.

BUG: 500052
FIXED-IN: 6.3.1


(cherry picked from commit 15a93e777be4afacff1683135f651518cb7afb9a)

Co-authored-by: Joshua Goins <josh@redstrate.com>