pager: remove onContainsMouseChanged, refresh model after desktopMouseArea.clicked
Due to refreshing the pager twice on mouse release and having two
different mouse drag tracking areas, these could get out of sync.
This caused the pagerModel refresh too early and lose any changes.
This changes it to only refresh after mouse is released, either when
dragging a window or clicking a desktop.

BUG:500236