folderview: Fix drag selection
There is code to handle clicking an icon when a selection is already
present. In this case we want to cancel our selection and select the
icon the mouse is over. What we don't want is for this to run when a
drag ends.

Using the cached selection is a good flag as it is immediately reset at
the end of the mouse up.

BUG: 499898


(cherry picked from commit 42f53e4618074d43d7956c6cbae0d5c4aa82825e)

Co-authored-by: David Edmundson <kde@davidedmundson.co.uk>