applets/kicker: focus the delegate in all cases
70f5b766 accidentally reverted the fix to focusing the delegate instead
of the column for the down arrow, leaving it only for the up arrow as
originally introduced in 8d10e42d.

This change adds it back. This is also a good chance to apply the fix to
left and right arrow keys, to make sure focusing items in other columns
also works immediately and to keep things consistent.

CCBUG: 499971