applets/taskmanager: scroll tooltip to active window
Currently, the tooltip shows the list of windows belonging to a grouped
task always starting at the first item. For some tasks involving many
windows the order is meaningful though, so it could be convenient to be
able to go to the previous/next window quickly. Similarly, if a user
scrolled to a window and selected it, then reconsidered, starting from
that window is likely more what the user expects, rather than having
to begin scrolling from the beginning again.

This change positions the ListView on showing the tooltip such that
if the grouped task contains the currently active window, it will be in
the visible range, and at the center if possible.

BUG: 499716