kcms/keyboard: Allow to edit the layout display name
The display name of the layouts can be customized, but in the UI it
was only possible to set it when adding a new one.

Expose the setting also in the layout list, in a non-intrussive way
by giving the appearance of a label, which can be edited when given
focus either by click or keyboard.

BUG: 500057
FIXED-IN: 6.4