applets/kickoff: don't unnecessarily look for icon name as local file
`Image{}` wants a file, not an icon name. When given one anyway, it
interprets the string as a local file, looks for it in the current
working dir, fails to find it, and then complains.

We know this will fail, so let's short-circuit the process by only
letting it look for strings that are clearly file paths.

BUG: 500373
FIXED-IN: 6.4.0