applets/window-list: show placeholder when empty
The window list in both forms (panel and desktop) does not handle the
case of having no windows open well; the desktop form is simply empty,
while the panel form has a completely empty menu active.

This change adds a Kirigami.PlaceHolder to the fullRepresentation and
sets a menu containing only a "No open windows" entry that does nothing
for the compactRepresentation. These are modeled on the way that the
task switchers handle this case. Along the way, we rename ids so that
"windowListView" refers to the ListView and not the component that wraps
it, because that is just confusing.

BUG: 481803

----


| before | after  |
| ------ | ------ |
![windowlist_before](/uploads/3056dff1bcbe2a1dce281a721f255a19/windowlist_before.png)| ![windowlist_after](/uploads/9221f6f0cdb998852e25c45aafd57464/windowlist_after.png)|

(Note that in the before screenshot, while the popup is empty and completely invisible, it's still modal, so the next click anywhere will do nothing except make the popup that's already not there even less there)