Port Timer objects away from adjustable-duration intervals where needed
A Timer's interval value should only use a standard duration unit that
adjusts with the user's chosen animation speed if it controls a visible
animation. Otherwise the interval will get scaled inappropriately and
weird things will happen.

For Timers where the duration should not scale, we should either use a
hardcoded value appropriate to the use case (if it needs to be short or
match something else) or else use `Kirigami.Units.humanMoment`, which
is 2 seconds long and targets this use case.

Several timers in this repo don't conform to those rules; port them to
the appropriate thing.

BUG: 500717
FIXED-IN: 6.3.3