applets/kicker: fix DashboardRepresentation warning
The ActionMenu component was changed to no longer inherit from Item.
While the use in cascading kicker was adapted to this, the one in
DashboardRepresentation was not. It stores the ActionMenu in an Item
property, and as the types do not match a warning is printed.

This change sets the type of the property to match what is actually
stored.


(cherry picked from commit 5e25d71aaf7bf6cd3920eca856e9e6fa1001a57c)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>