applets/taskmanager: handle null playerData when switching tooltip
The tooltips conditionally load the player controls depending on whether
the playerData property is set, and the controls also access this
property. This leads to a race condition when changing the hovered app
while the tooltip is showing and the tooltip goes from actual playerData
to null: if the Loader stops noticing the changed property first all is
well, but if the PlayerController notices it first, it spews a slew of
TypeErrors before being unloaded.

Instead, we guard against null playerData and set appropriate values in
those cases so the unloading of the PlayerController can always happen
gracefully.