applets/keyboardlayout: fix warning on startup
The applet sets the font point size to the applet height. While the
applet is loading, its height is 0, which causes QFont to complain as
it prefers its point sizes strictly greater than zero.

Instead, just set it to something reasonable if falsy - taking the
Kirigami default point size here because it sounds reasonable; if the
applet is displayed it'll have a height anyway so the actual value
doesn't really matter.