kcms/runners: Add power/session actions to favorite runners list
Right now we have a problem that when people search for "shut down" and
"reboot" and the like, the Desktop Sessions KCM appears first. This is
currently expected because the System Settings runner is a default
favorite runner, and the power/session runners are not. Logical, but not
ideal.

There are three ways we can fix this:
1. Add the runners for power and session actions to the default favorite
   runners list.
2. Remove a bunch of keywords from the desktop sessions KCM so it
   doesn't match those kinds of searches anymore.
3. Remove all default favorite runners and hope the internal weightings
   are good enough to solve the problem.

This commit implements option 1, which is the least invasive change
given the current implementations.

BUG: 474981
FIXED-IN: 6.4.0