containments/desktop: resolve binding loop
The assignment of minimum and preferred sizes on the FolderView has
binding loops, which result from intermediate values as the other
bindings are filled.

This change resolves these loops by setting the bindings using
Qt.callLater(); when the main loop is clear and the other bindings have
settled, the proper sizes can be determined without causing loops.

BUG: 483859