desktoppackage/configuration: handle loading more gracefully
ConfigurationContainmentAppearance checks whether the containment
switching warning needs to be shown or hidden when the value is changed.
This also happens when the plugin is loaded, and at that point the
stackview does not have a currentItem yet, so the check for whether to
hide it fails.

Instead use optional chaining.