Use entire screen to place sidebars when not in edit mode
We previously made sure that the sidebar would avoid panels
to allow for easier drag and drop from the widget sidebar,
since that's only supposed to happen in edit mode we only
consider panels in edit mode, and instead use the entire
screen otherwise. If we try to avoid panels when outside
edit mode, we won't be able to know if dodge window panels
are visible, and if they're not the sidebar will be visually
broken.

BUG:480890


(cherry picked from commit 12e658537ee6cb531176b688cdc3a43642b44d0c)

Co-authored-by: Niccolò Venerandi <niccolo@venerandi.com>