containments/desktop: only bind AppletsLayout if containment
The desktop containment is used both for the desktop and for other
FolderViews on the panel or desktop. Only the full desktop folderview
can contain further applets; panel applets or extra folder views on the
desktop are not containments. However, they still try to set themselves
up with a ContainmentLayoutManager.AppletsLayout, which fails with a
warning because they are not containments.

Instead, only set the AppletsLayout containment property if it's
actually a containment. This is effectively what happens anyway if the
assignment fails, but skips the warning.