desktoppackage/WidgetExplorer: avoid scrollbar binding loop
The scroll bar visibility can cause a binding loop as it becoming
visible can change the word wrap of the contents, which may causes the
size to increase, which triggers re-evaluation of the visibility. This
is jsut a temporary problem, as the size can't shrink due to the
scrollbar becoming visible, and there's only one vertical scrollbar.

Instead, use a delayed binding to give it some extra time. This is
similar to how this was handled in the notes applet.