desktoppackage/configuration: simplify containment config loading
The two global containment configuration pages (Wallpaper and Actions)
still use the old loading method, which is essentially a copy of the
regular method that also wraps them inside a page and adds an extra
configuration flag. This makes the whole structure more complicated and
harder to modify.

Instead, convert the two pages to use KCMUtils base classes and load
them the usual way.

BUG: 501584