containments/desktopbase configuration dialog pages on KCMUtils
The Folder View configuration pages are still loaded as containments (or
in the case of panel Folder Views, are incorrectly loaded with the
default config page loader, which causes them to not have page headers).
With the containment-specific loaders gone, these need to be switched to
KCMUtils components rather than arbitrary items as their base.

This change switches them to SimpleKCM or ScrollViewKCM as appropriate.