desktoppackage/configuration: only set necessary wallpaper config
ConfigurationContainmentAppearance sets all existing configuration
values for the loaded wallpaper plugin, but many plugins do not make use
of the default value properties that kconfig provides.

Instead of setting all of them, load the plugin and only set the ones
that are actually defined in the configuration page. The configDialog
property remains as an initial property even though most plugins do not
use it, as additional work is required on the affected plugins to have
it loadable later.