applets/taskmanager: fix sorting configuration combobox
The sorting combobox uses a model as the index of the entry does
not match the enum values used in the backend (one of the entries is
currently not exposed). In that case, however, the value cannot be set
through a property alias - currentValue only works for reading values,
not for setting them.

Instead, use a real property and read/set it as in the qml documentation
on comboboxes with data models.

BUG: 502247