desktoppackage/AppletConfiguration: consider more keys unchanged
Several configuration pages use qml lists of (usually) strings as data
stored under a configuration key. These are considered objects by the
qml engine, and will not evaluate the same even if they have the same
contents.

Instead, also require that the data has to evaluate to the same string
representation to be considered modified.