applets/kicker: reset custom icon when cleared
Kicker's configuration page has a somewhat complex handling of the
custom icon, which is spread across three properties. When the custom
icon is cleared, the property that governs which of the other properties
is used is set. but the actual icon is not cleared. This means that
setting and resetting the icon will still prompt the user to save the
changes to the (now unused) property. This would be meaningful if the
custom icon could be reactivated, but the only way to enable it again is
to set it again, at which point the stored value would be overwritten
anyway.

Instead, remove the value of the custom property as well if the icon is
cleared.