applets/kicker: handle icon dialog being accepted
Kicker's custom icon button handles the icon dialog's iconName property
being changed, and stores the result in its own configuration. This will
fail sometimes: if the user sets a custom icon, clears it using the icon
button's "Clear Icon" function, and sets the same custom icon again, the
icon will not be changed as the icon dialog's property is unchanged by
selecting the same icon again.

Instead, we listen to the icon dialog's accepted signal, and only set
the custom icon if an icon was actually chosen.