desktoppackage/AppletConfiguration: track unsavedChanges property
b542ac85 introduced support for a property unsavedChanges that applet
configuration pages can use to communicate to the dialog that some
configuration properties that are independently managed are presently
changed and were not yet saved. This allows extending the improved check
for unsaved changes to parts of the configuration that cannot be tracked
by the usual mechanism for tracking cfg_ keys. While the configuration
dialog does take this property into account, it does not independently
track it, so the interface might not update if its value changes.

This change adds a handler that listens to changes in this property.