desktoppackage/AppletConfiguration: fix onSettingValueChanged
onSettingValueChanged is an internal signal that was used for recursive
embedding (e.g. for containment configuration when they used their own
loading mechanism) that basically did the same thing as the existing
configurationChanged signal: mark the configuration as changed so users
will be prompted to save or discard these changes. It shouldn't be
necessary anymore, but we can't rule out that third-party plasmoids may
exist that use it instead of configurationChanged, so we should not
remove it at this point. It does not integrate properly with the new
dynamic handling of tracking configuration changes in b542ac85.

This change makes onSettingValueChanged work like configurationChanged()
does, and adds a comment indicating that it should not be used in new
code and may be removed in Plasma 7.


Co-authored-by: Nate Graham <nate@kde.org>