applets/kimpanel: fix show/hide
Kimpanel stores the menu separator (a MenuItem) in an Item-typed
property. Confusingly enough, MenuItem is not actually an Item subclass,
but a direct QtObject. This not only causes a warning, if the Hide
context menu action is used, it's not possible to undo this as the menu
fails in that case due to the property being null.

Instead, make it a property with type MenuItem, which fixes the warning
and restores functionality.