desktoppackage/configuration: don't force breadcrumb header
The applet configuration dialog pages now have to be based on KCMUtils
components, but the page header style is forced to be Breadcrumb. This
makes some of the typical KCM patterns (such as a GridView KCM with
Add or KNS buttons in the header) impossible.

Instead, use Auto style, this will allow use of actions in the header on
desktop and should also work on mobile as KCMs do now.