containments/desktop: set implicitHeight for preview config dialog
The preview configuration (sub)dialog contains only a ListView as its
body, which has no implicitSize and so causes a binding loop with
Dialog's centering logic. In addition, the dialog's size extends until
the bottom of the dialog, rather than be centered.

Instead, set its implicitHeight to 80% of the parent's height, similar
to how this was handled in e.g. the Add Dictionary dialog in the
Dictionary applet's configuration or the Add Language dialog in
kcm_regionandlang.