desktoppackage/configuration: track unsaved changes for Shortcuts page
The Shortcuts configuration page currently notifies the configuration
dialog through the configurationChanged() signal that there are unsaved
changes; this is not responsive to the user restoring the original
configuration manually.

Instead, set the unsavedChanges property as appropriate. Along the way,
we switch to the keySequenceModified signal, as captureFinished is
deprecated.